+++ packages/paste-core/components/in-page-navigation/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @twilio-paste/in-page-navigation
 
+## 4.3.3
+
+### Patch Changes
+
+- [`e9586bdf8`](https://github.com/twilio-labs/paste/commit/e9586bdf82321ec1815a6df3752f50a20677fe4f) [#4143](https://github.com/twilio-labs/paste/pull/4143) Thanks [@krisantrobus](https://github.com/krisantrobus)! - [InPageNavigation] fix issue with currently selected item causing vertical scroll
+
 ## 4.3.2
 
 ### Patch Changes